// app/src/generated/types/bakeEntry.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getAddressDecoder,
  getAddressEncoder,
  getI64Decoder,
  getI64Encoder,
  getStructDecoder,
  getStructEncoder,
  getU64Decoder,
  getU64Encoder,
  type Address,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export type BakeEntry = { baker: Address; at: bigint; potAfter: bigint };

export type BakeEntryArgs = {
  baker: Address;
  at: number | bigint;
  potAfter: number | bigint;
};

export function getBakeEntryEncoder(): FixedSizeEncoder<BakeEntryArgs> {
  return getStructEncoder([
    ["baker", getAddressEncoder()],
    ["at", getI64Encoder()],
    ["potAfter", getU64Encoder()],
  ]);
}

export function getBakeEntryDecoder(): FixedSizeDecoder<BakeEntry> {
  return getStructDecoder([
    ["baker", getAddressDecoder()],
    ["at", getI64Decoder()],
    ["potAfter", getU64Decoder()],
  ]);
}

export function getBakeEntryCodec(): FixedSizeCodec<BakeEntryArgs, BakeEntry> {
  return combineCodec(getBakeEntryEncoder(), getBakeEntryDecoder());
}