// app/src/generated/types/winnerEntry.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getAddressDecoder,
  getAddressEncoder,
  getI64Decoder,
  getI64Encoder,
  getStructDecoder,
  getStructEncoder,
  getU32Decoder,
  getU32Encoder,
  getU64Decoder,
  getU64Encoder,
  type Address,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export type WinnerEntry = {
  round: bigint;
  winner: Address;
  prize: bigint;
  bakes: number;
  at: bigint;
};

export type WinnerEntryArgs = {
  round: number | bigint;
  winner: Address;
  prize: number | bigint;
  bakes: number;
  at: number | bigint;
};

export function getWinnerEntryEncoder(): FixedSizeEncoder<WinnerEntryArgs> {
  return getStructEncoder([
    ["round", getU64Encoder()],
    ["winner", getAddressEncoder()],
    ["prize", getU64Encoder()],
    ["bakes", getU32Encoder()],
    ["at", getI64Encoder()],
  ]);
}

export function getWinnerEntryDecoder(): FixedSizeDecoder<WinnerEntry> {
  return getStructDecoder([
    ["round", getU64Decoder()],
    ["winner", getAddressDecoder()],
    ["prize", getU64Decoder()],
    ["bakes", getU32Decoder()],
    ["at", getI64Decoder()],
  ]);
}

export function getWinnerEntryCodec(): FixedSizeCodec<
  WinnerEntryArgs,
  WinnerEntry
> {
  return combineCodec(getWinnerEntryEncoder(), getWinnerEntryDecoder());
}